# internal/adapter/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "# Introduction\nThis is documentation for Gathering App API",
        "title": "Gathering App API",
        "contact": {},
        "version": "v1.0.0"
    },
    "paths": {
        "/gatherings": {
            "get": {
                "description": "Get Gatherings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gathering"
                ],
                "summary": "Get Gatherings",
                "responses": {
                    "200": {
                        "description": "Gathering",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/helpers.ResponsePayload"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "data": {
                                                "$ref": "#/definitions/swaggermodel.Gathering"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Gathering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gathering"
                ],
                "summary": "Create Gathering",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.Gathering"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gathering",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponsePayload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/swaggermodel.Gathering"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/gatherings/{id}": {
            "get": {
                "description": "Get Gathering By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gathering"
                ],
                "summary": "Get Gathering By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gathering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gathering",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponsePayload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/swaggermodel.Gathering"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update Gathering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gathering"
                ],
                "summary": "Update Gathering",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gathering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.UpdateGathering"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gathering",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponsePayload"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Gathering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gathering"
                ],
                "summary": "Delete Gathering",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gathering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gathering",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponsePayload"
                        }
                    }
                }
            }
        },
        "/invitations": {
            "get": {
                "description": "Get Invitations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitation"
                ],
                "summary": "Get Invitations",
                "responses": {
                    "200": {
                        "description": "Invitation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/helpers.ResponsePayload"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "data": {
                                                "$ref": "#/definitions/swaggermodel.Invitation"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create Invitation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitation"
                ],
                "summary": "Create Invitation",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swaggermodel.Invitation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitation",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponsePayload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/swaggermodel.Invitation"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/invitations/{id}": {
            "get": {
                "description": "Get Invitation By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitation"
                ],
                "summary": "Get Invitation By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invitation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitation",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponsePayload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/swaggermodel.Invitation"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helpers.ResponsePayload": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "swaggermodel.Gathering": {
            "type": "object",
            "required": [
                "creator",
                "location",
                "name",
                "scheduled_at",
                "type"
            ],
            "properties": {
                "attendees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/swaggermodel.MemberPayload"
                    }
                },
                "creator": {
                    "$ref": "#/definitions/swaggermodel.MemberPayload"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string",
                    "example": "gathering street"
                },
                "name": {
                    "type": "string",
                    "example": "Gathering Name"
                },
                "scheduled_at": {
                    "description": "Date using (YYYY-MM-DD MM:SS) format",
                    "type": "string",
                    "example": "2023-10-06 04:53"
                },
                "type": {
                    "description": "Gathering type\n* 0 -\u003e Private\n* 1 -\u003e Public",
                    "allOf": [
                        {
                            "$ref": "#/definitions/valueobject.GatheringType"
                        }
                    ],
                    "example": 1
                }
            }
        },
...
        "swaggermodel.MemberPayload": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}